requirements.txt
# Core dependencies
numpy==1.24.3
resampy==0.4.3
librosa==0.10.0
s3tokenizer
torch==2.1.0
torchaudio==2.1.0
transformers==4.46.3
diffusers==0.29.0
resemble-perth==1.0.1
omegaconf==2.3.0
conformer==0.3.2
spacy>=3.4.0

# API dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0

# Audio processing
soundfile>=0.12.0

# RunPod integration
runpod==1.3.0
requests>=2.26.0

# UI
gradio==3.50.2 